move waymax to gocarx
