+++ waymax/metrics/__init__.py
@@ -25,6 +25,6 @@ from waymax.metrics.roadgraph import OffroadMetric
 from waymax.metrics.roadgraph import WrongWayMetric
 from waymax.metrics.route import OffRouteMetric
 from waymax.metrics.route import ProgressionMetric
-from waymax.metrics.gokart_metric import GokartProgressMetric
-from waymax.metrics.gokart_metric import GokartOrientationMetric
-from waymax.metrics.gokart_metric import GokartOffroadMetric
+from waymax.metrics.gokart_progress import GokartProgressMetric
+from waymax.metrics.gokart_orientation import GokartOrientationMetric
+from waymax.metrics.gokart_offroad import GokartOffroadMetric
